requirements.txt
asgiref==3.2.10
astroid==2.2.5
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.0
bleach==3.0.2
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlover==0.2.1
cufflinks==0.14.6
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
djangorestframework==3.11.0
easygui==0.98.1
entrypoints==0.2.3
et-xmlfile==1.0.1
future==0.17.1
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
importlib-metadata==1.7.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.15
jdcal==1.4.1
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numpy==1.15.4
openpyxl==3.0.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==7.1.2
plotly==3.4.2
pluggy==0.9.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyotp==2.2.7
pyparsing==2.3.0
PyQt5==5.12
PyQt5-sip==4.19.14
python-dateutil==2.7.5
python-jsonrpc-server==0.1.2
python-language-server==0.25.0
pytz==2018.7
pywinpty==0.5.5
pyzmq==17.1.2
qtgui==0.0.1
requests==2.21.0
retrying==1.3.3
rope==0.14.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
soupsieve==1.6
sqlparse==0.3.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
yapf==0.26.0
zerosms==0.0.1
zipp==3.1.0
